x86_64: Moved pagetable functions into separte mod
